Move GuzzleCompat to Internal namespace
