+++ src/Internal/GuzzleCompat.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bugsnag;
+namespace Bugsnag\Internal;
 
 use GuzzleHttp;
